tests/submission/Infraware/Forms/contents/Forms/legend_form.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>Forms</title>
  <script src="../../resources/testharness.js"></script>
  <script src="../../resources/testharnessreport.js"></script>
 </head>
 <body>
	<p>
		<h3>legend_form</h3>
	</p>
	
	<hr>

	<div id="log"></div>

	<form method="post"
		  enctype="application/x-www-form-urlencoded"
		  action=""
		  id="input_form"
		  name="input_form">
		<fieldset id="input_field">
			<legend id='input_legend'></legend>
 		</fieldset>
	</form>
	<script>

		var legend = document.getElementById("input_legend");

		try
		{
			var form = legend.form;

			test(function() {
				assert_equals(form.name, "input_form", "form attribute is not correct.");
			});		
		}
		catch (e)
		{
			test(function() {
				assert_unreached("Error is raised.");
			});
		}

	</script>

 </body>
</html>
